fix press within popover bug
